requirements.txt
numpy
sentence_transformers
faiss-cpu
langchain
langchain_openai
fastcluster
FlagEmbedding